Separate PreVoteCollector out as a top-level class
