+++ server/src/test/java/org/elasticsearch/cluster/coordination/ElectionSchedulerTests.java
@@ -46,8 +46,8 @@ import static java.util.Collections.singleton;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_BACK_OFF_TIME_SETTING;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_MAX_TIMEOUT_SETTING;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.ELECTION_MIN_TIMEOUT_SETTING;
-import static org.elasticsearch.cluster.coordination.ElectionScheduler.REQUEST_PRE_VOTE_ACTION_NAME;
 import static org.elasticsearch.cluster.coordination.ElectionScheduler.validationExceptionMessage;
+import static org.elasticsearch.cluster.coordination.PreVoteCollector.REQUEST_PRE_VOTE_ACTION_NAME;
 import static org.elasticsearch.node.Node.NODE_NAME_SETTING;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
